requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
dj-rest-auth==6.0.0
Django==4.2.15
django-allauth==64.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
executing==2.0.1
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.7
mongomock==4.1.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pymongo==3.12.3
python-decouple==3.8
pytz==2024.1
sentinels==1.0.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13